guardar contacto y carrousel de imagenes

+++ src/main/java/com/mycompany/proyecto_estructuras/CreateContactController.java
@@ -1,12 +1,18 @@
 package com.mycompany.proyecto_estructuras;
 
-import Direcciones.Direccion;
-import Fechas.Fecha;
+import Direcciones.DireccionCb;
+import Fechas.FechaCb;
+import Logica.Archivos;
+import Logica.Direccion;
 import Logica.DoubleLinkedList;
+import Logica.Fecha;
 import Logica.Persona;
+import Logica.RedSocial;
 import Logica.Telefono;
+import Logica.Usuario;
 import Prefijos.PrefijoPais;
 import Social_Media.RedesSociales;
+import static com.mycompany.proyecto_estructuras.App.listaUsuarios;
 import java.io.File;
 import java.io.IOException;
 import javafx.event.ActionEvent;
@@ -23,6 +29,9 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.nio.file.StandardCopyOption;
+import java.time.LocalDate;
+import java.time.format.DateTimeFormatter;
+import java.util.ListIterator;
 import java.util.ResourceBundle;
 import javafx.event.Event;
 import javafx.scene.Node;
@@ -71,9 +80,11 @@ public class CreateContactController implements Initializable {
     @FXML
     private ComboBox<RedesSociales> comboBoxRedes;
     @FXML
-    private ComboBox<Direccion> comboDirections;
+    private ComboBox<DireccionCb> comboDirecciones;
     @FXML
-    private ComboBox<Fecha> comboFechas;
+    private ComboBox<FechaCb> comboFechas;
+    private DoubleLinkedList<String> fotos = new DoubleLinkedList<>();
+    private ListIterator<String> iterator = fotos.listIterator();
 
     @FXML
     private void cambiarTipo(ActionEvent event) {
@@ -102,12 +113,12 @@ public class CreateContactController implements Initializable {
 
     @FXML
     private void handleComboBoxDirections(Event event) {
-        Direccion.configurarComboBoxConDirecciones(comboDirections);
+        DireccionCb.configurarComboBoxConDirecciones(comboDirecciones);
     }
 
     @FXML
     private void handleComboBoxDates(Event event) {
-        Fecha.configurarComboBoxConFechas(comboFechas);
+        FechaCb.configurarComboBoxConFechas(comboFechas);
     }
 
     @FXML
@@ -116,16 +127,16 @@ public class CreateContactController implements Initializable {
 
         VBox vbox = (VBox) ((HBox) sourceButton.getParent().getParent()).getChildren().get(0);
         agregarEliminar(sourceButton, vbox);
-        if (vbox == cajaTelefonos ) {
-            agregarTextField(vbox,"telefonos");
+        if (vbox == cajaTelefonos) {
+            agregarTextField(vbox, "telefonos");
         } else if (vbox == cajaEmails) {
             agregarTextFieldEnHBox(vbox);
         } else if (vbox == cajaDirecciones) {
-            agregarTextField(vbox,"direcciones");
+            agregarTextField(vbox, "direcciones");
         } else if (vbox == cajaFechas) {
             agregarTextFieldYDatePickerEnHBox(vbox);
-        } else if (vbox == cajaRedes){
-            agregarTextField(vbox,"redes");
+        } else if (vbox == cajaRedes) {
+            agregarTextField(vbox, "redes");
         }
 
     }
@@ -135,9 +146,9 @@ public class CreateContactController implements Initializable {
         hBox.setSpacing(20);
         ComboBox<PrefijoPais> comboBox1 = new ComboBox<PrefijoPais>();
         ComboBox<RedesSociales> comboBox2 = new ComboBox<RedesSociales>();
-        ComboBox<Direccion> comboBox3 = new ComboBox<Direccion>();
+        ComboBox<DireccionCb> comboBox3 = new ComboBox<DireccionCb>();
 
-        ComboBox<Fecha> comboBox4 = new ComboBox<Fecha>();
+        ComboBox<FechaCb> comboBox4 = new ComboBox<FechaCb>();
         comboBox1.setMinWidth(150);
         comboBox2.setMinWidth(150);
         comboBox3.setMinWidth(150);
@@ -151,12 +162,12 @@ public class CreateContactController implements Initializable {
         } else if (tipo.equalsIgnoreCase("redes")) {
             RedesSociales.configurarComboBoxConRedes(comboBox2);
             hBox.getChildren().add(comboBox2);
-            
+
         } else if (tipo.equalsIgnoreCase("direcciones")) {
-            Direccion.configurarComboBoxConDirecciones(comboBox3);
+            DireccionCb.configurarComboBoxConDirecciones(comboBox3);
             hBox.getChildren().add(comboBox3);
 
-        } 
+        }
 
         TextField textField = new TextField();
         hBox.getChildren().add(textField);
@@ -165,7 +176,7 @@ public class CreateContactController implements Initializable {
 
     @FXML
     private void handleBtnFotoClick(ActionEvent event) throws IOException {
-        Stage stage = (Stage) ((Node) event.getSource()).getScene().getWindow(); // Obtiene la ventana actual
+        Stage stage = (Stage) ((Node) event.getSource()).getScene().getWindow();
 
         FileChooser fileChooser = new FileChooser();
         fileChooser.setTitle("Selecciona una imagen");
@@ -175,10 +186,26 @@ public class CreateContactController implements Initializable {
         if (selectedFile != null) {
             // Asumiendo que "Img" es una carpeta en el mismo directorio que el directorio de salida del proyecto.
             Path destPath = Paths.get("Img" + File.separator + selectedFile.getName());
-            Files.copy(selectedFile.toPath(), destPath, StandardCopyOption.REPLACE_EXISTING);
-            // Si deseas reflejar la imagen en la UI:
-            Image image = new Image(destPath.toUri().toString());
-            ImgFotoPersona.setImage(image); // Asegúrate de tener un ImageView en tu FXML para mostrar la imagen.
+
+            // Verificar si el archivo ya existe en el directorio
+            if (Files.exists(destPath)) {
+                // El archivo ya existe, puedes mostrar un mensaje o realizar alguna acción.
+                System.out.println("La imagen ya existe en el directorio.");
+            } else {
+                // El archivo no existe, copiarlo y agregarlo a la lista
+                Files.copy(selectedFile.toPath(), destPath, StandardCopyOption.REPLACE_EXISTING);
+
+                // Reflejar la imagen en la UI
+                String rutaImagen = destPath.toUri().toString();
+                Image image = new Image(rutaImagen);
+                ImgFotoPersona.setImage(image);
+                fotos.add(rutaImagen);
+                System.out.println("se agregó");
+                if (iterator.hasNext()) {
+                    iterator.next();
+                }
+
+            }
         }
     }
 
@@ -200,8 +227,8 @@ public class CreateContactController implements Initializable {
     private void agregarTextFieldYDatePickerEnHBox(VBox parentVBox) {
         HBox hBox = new HBox();
         hBox.setSpacing(10);
-        ComboBox<Fecha> cb = new ComboBox();
-        Fecha.configurarComboBoxConFechas(cb);
+        ComboBox<FechaCb> cb = new ComboBox();
+        FechaCb.configurarComboBoxConFechas(cb);
         DatePicker datePicker = new DatePicker();
         hBox.getChildren().addAll(cb, datePicker);
         parentVBox.getChildren().add(hBox);
@@ -246,18 +273,21 @@ public class CreateContactController implements Initializable {
             System.out.println("redesSociales: " + redesSociales);
             System.out.println("direcciones: " + direcciones);
             System.out.println("fechasRelevantes: " + fechasRelevantes);
-            DoubleLinkedList<String[]> lldirecciones = new DoubleLinkedList();
+            DoubleLinkedList<Direccion> lldirecciones = new DoubleLinkedList();
             DoubleLinkedList<String> llemails = new DoubleLinkedList();
-            DoubleLinkedList<String> llredes = new DoubleLinkedList();
+            DoubleLinkedList<RedSocial> llredes = new DoubleLinkedList();
             DoubleLinkedList<String> llfotos = new DoubleLinkedList();
-            DoubleLinkedList<String[]> llfechas = new DoubleLinkedList();
+            DoubleLinkedList<Fecha> llfechas = new DoubleLinkedList();
             DoubleLinkedList<Telefono> lltelefonos = new DoubleLinkedList();
+            llfotos.addAll(fotos);
             //añadir telefonos
             String[] telefonosArray = telefonos.split("\\|");
             String[] emailsArray = emails.split("\\|");
             String[] redesSocialesArray = redesSociales.split("\\|");
             String[] direccionesArray = direcciones.split("\\|");
             String[] fechasRelevantesArray = fechasRelevantes.split("\\,");
+            DateTimeFormatter formateo = DateTimeFormatter.ofPattern("yyyy-MM-dd");
+            LocalDate cumpleaños = null;
             for (String algo : telefonosArray) {
                 if (!algo.isEmpty()) {
                     String[] telefonoArray = algo.split("[()]");
@@ -269,15 +299,26 @@ public class CreateContactController implements Initializable {
                 llemails.add(algo);
             }
             for (String algo : redesSocialesArray) {
-                llredes.add(algo);
+                String[] red = algo.split(" ");
+                RedSocial rs = new RedSocial(red[0], red[1]);
+                llredes.add(rs);
             }
             for (String algo : direccionesArray) {
                 String[] direccion = algo.split(" ");
-                lldirecciones.add(direccion);
+                Direccion d = new Direccion(direccion[0], direccion[1]);
+                lldirecciones.add(d);
             }
             for (String algo : fechasRelevantesArray) {
-                String[] fecha = algo.split("\\|");
-                llfechas.add(fecha);
+                String[] fecha = algo.split(" ");
+                if (fecha[0].equals("Cumpleaños")) {
+                    LocalDate fechaComoLocalDate = LocalDate.parse(fecha[1], formateo);
+                    cumpleaños = fechaComoLocalDate;
+                }
+                LocalDate localD = LocalDate.parse(fecha[1], formateo);
+                Fecha f = new Fecha(fecha[0], localD);
+
+                llfechas.add(f);
+
             }
 
 //            Persona(String apellido , String genero, String fechaNacimiento
@@ -287,8 +328,17 @@ public class CreateContactController implements Initializable {
 //            , DoubleLinkedList<String> fotos, DoubleLinkedList<String[]> fechas
 //            , DoubleLinkedList<String[]> telefonos
 //            ) 
-            Persona contacto = new Persona(apellidos, genero, fechasRelevantes, ocupacion, Nacionalidad, nombres, lldirecciones, llemails, llredes, llfotos, llfechas, lltelefonos);
+            Persona contacto = new Persona(apellidos, genero, cumpleaños, ocupacion, Nacionalidad, nombres, lldirecciones, llemails, llredes, llfotos, llfechas, lltelefonos);
             System.out.println("Contacto: " + contacto);
+            App.usuario.getContactos().add(contacto);
+            for (Usuario usuario : App.listaUsuarios) {
+                if (usuario.equals(App.usuario)) {
+                    usuario.getContactos().add(contacto);
+                    break;
+                }
+            }
+            Archivos.serializarListaUsuarios(App.listaUsuarios, "usuarios.ser");
+
         }
     }
 
@@ -345,4 +395,52 @@ public class CreateContactController implements Initializable {
             nodoPadre.getChildren().add(nuevoBoton);
         }
     }
+
+    @FXML
+    private void handleBtnSiguienteClick(ActionEvent event) {
+        if (!fotos.isEmpty()) {
+            if (iterator.hasNext()) {
+                System.out.println("pa delante");
+                String siguienteFoto = iterator.next();
+                mostrarImagenActual(siguienteFoto);
+            } else {
+                // Volvemos al principio de la lista
+                iterator = fotos.listIterator();
+                if (iterator.hasNext()) {
+                    String siguienteFoto = iterator.next();
+                    mostrarImagenActual(siguienteFoto);
+                }
+            }
+        } else {
+            System.out.println("bug");
+        }
+    }
+
+    @FXML
+    private void handleBtnAnteriorClick(ActionEvent event) {
+        if (!fotos.isEmpty()) {
+            System.out.println("para atras");
+            if (iterator.hasPrevious()) {
+                String anteriorFoto = iterator.previous();
+                mostrarImagenActual(anteriorFoto);
+            } else {
+                // Vamos al final de la lista
+                while (iterator.hasNext()) {
+                    iterator.next();
+                }
+                if (iterator.hasPrevious()) {
+                    String anteriorFoto = iterator.previous();
+                    mostrarImagenActual(anteriorFoto);
+                }
+            }
+        } else {
+            System.out.println("bug 2");
+        }
+    }
+
+    private void mostrarImagenActual(String rutaImagen) {
+        Image image = new Image(rutaImagen);
+        ImgFotoPersona.setImage(image);
+    }
+
 }
